Move error.h to not be compiler specific.
